pyproject.toml
[project]
name = "koltin_challenge"
requires-python = ">=3.9,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.11.16",
    "dagster-dbt>=0.27.16",
    "dbt-duckdb>=1.9.6",
    "dbt-snowflake>=1.10.2",
    "pandas>=2.3.3",
    "snowflake-sqlalchemy>=1.7.7",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "koltin_challenge"
registry_modules = [
    "koltin_challenge.components.*",
]


